+++ ts-client/src/test/rebalance_with_strategy.test.ts
@@ -238,7 +238,7 @@ describe("Rebalance with strategy", () => {
           index: index.toNumber(),
           binStep: DEFAULT_BIN_STEP.toNumber(),
           baseFactor: DEFAULT_BASE_FACTOR_2.toNumber(),
-          functionType: FunctionType.LiquidityMining,
+          concreteFunctionType: FunctionType.LiquidityMining,
           filterPeriod: 30,
           decayPeriod: 600,
           reductionFactor: 5000,
@@ -246,6 +246,7 @@ describe("Rebalance with strategy", () => {
           protocolShare: 0,
           maxVolatilityAccumulator: 350000,
           baseFeePowerFactor: 0,
+          collectFeeMode: 0,
         })
         .accountsPartial({
           signer: keypair.publicKey,